Eliminate duplicate code for parsing timescale from init segments
