+++ lib/media/timestamp_offset_corrector.js
@@ -12,6 +12,7 @@ goog.require('shaka.util.FakeEvent');
 goog.require('shaka.util.ManifestParserUtils');
 goog.require('shaka.util.Mp4BoxParsers');
 goog.require('shaka.util.Mp4Parser');
+goog.require('shaka.util.Mp4SegmentParsers');
 
 /**
  * Corrects the timestampOffset in SegmentReference(s) by
@@ -223,43 +224,8 @@ shaka.media.TimestampOffsetCorrector = class {
       return;
     }
 
-    const Mp4Parser = shaka.util.Mp4Parser;
-    const trackIds = [];
-    const timescales = [];
-
-    new Mp4Parser()
-        .box('moov', Mp4Parser.children)
-        .box('mvex', Mp4Parser.children)
-        .box('trak', Mp4Parser.children)
-        .fullBox('tkhd', (box) => {
-          if (box.version !== null) {
-            const parsedTKHDBox = shaka.util.Mp4BoxParsers.parseTKHD(
-                box.reader, box.version);
-            trackIds.push(parsedTKHDBox.trackId);
-          }
-        })
-        .box('mdia', Mp4Parser.children)
-        .fullBox('mdhd', (box) => {
-          if (box.version !== null) {
-            const parsedMDHDBox = shaka.util.Mp4BoxParsers.parseMDHD(
-                box.reader, box.version);
-            timescales.push(parsedMDHDBox.timescale);
-          }
-        })
-        .parse(initSegment, /* partialOkay= */ true);
-
-    /** @type {Map<number, number>} */
-    const trackTimescales = new Map();
-
-    // Populate the map from trackId to timescale
-    trackIds.forEach((trackId, idx) => {
-      trackTimescales.set(trackId, timescales[idx]);
-    });
-
-    shaka.log.v2('parseTimescalesFromInitSegment() ',
-        'contentType', contentType,
-        'trackIds', JSON.stringify(trackIds),
-        'timescales', JSON.stringify(timescales));
+    const trackTimescales =
+        shaka.util.Mp4SegmentParsers.parseTrackTimescales(initSegment);
 
     const mediaData = this.getMediaData_(contentType);
     mediaData.trackTimescales = trackTimescales;
